requirements.txt
quart
quart-cors
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pypdf==3.15.5
langchain==0.0.276
openai==0.27.10
python-dotenv==1.0.0
tiktoken==0.4.0
faiss-cpu==1.7.4